moved data/old_data.h to data/old/old_data.h for better organization, moved currently unused legion barrier test from test/barrier to test/legion/barrier
